moved write staff to StaffRW
