+++ src/engraving/infrastructure/io/xmlwriter.cpp
@@ -74,7 +74,7 @@ void XmlWriter::putLevel()
 //   header
 //---------------------------------------------------------
 
-void XmlWriter::header()
+void XmlWriter::writeHeader()
 {
     *this << "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n";
 }
