+++ src/engraving/libmscore/chordlist.cpp
@@ -1911,7 +1911,7 @@ bool ChordList::write(const QString& name) const
 bool ChordList::write(QIODevice* device) const
 {
     XmlWriter xml(0, device);
-    xml.header();
+    xml.writeHeader();
     xml.startObject("museScore version=\"" MSC_VERSION "\"");
 
     write(xml);
