+++ src/engraving/style/style.cpp
@@ -340,7 +340,7 @@ void MStyle::read(XmlReader& e, compat::ReadChordListHook* readChordListHook)
 bool MStyle::write(QIODevice* device)
 {
     XmlWriter xml(nullptr, device);
-    xml.header();
+    xml.writeHeader();
     xml.startObject("museScore version=\"" MSC_VERSION "\"");
     save(xml, false);
     xml.endObject();
