+++ src/engraving/tests/testbase.cpp
@@ -65,7 +65,7 @@ EngravingItem* MTest::writeReadElement(EngravingItem* element)
     QBuffer buffer;
     buffer.open(QIODevice::WriteOnly);
     XmlWriter xml(element->score(), &buffer);
-    xml.header();
+    xml.writeHeader();
     element->write(xml);
     buffer.close();
