+++ src/palette/internal/palette.cpp
@@ -442,7 +442,7 @@ bool Palette::writeToFile(const QString& p) const
     QBuffer cbuf;
     cbuf.open(QIODevice::ReadWrite);
     XmlWriter xml(gpaletteScore, &cbuf);
-    xml.header();
+    xml.writeHeader();
     xml.startObject("container");
     xml.startObject("rootfiles");
     xml.startObject(QString("rootfile full-path=\"%1\"").arg(XmlWriter::xmlString("palette.xml")));
@@ -467,7 +467,7 @@ bool Palette::writeToFile(const QString& p) const
         QBuffer cbuf1;
         cbuf1.open(QIODevice::ReadWrite);
         XmlWriter xml1(gpaletteScore, &cbuf1);
-        xml1.header();
+        xml1.writeHeader();
         xml1.startObject("museScore version=\"" MSC_VERSION "\"");
         write(xml1);
         xml1.endObject();
